Organization-ID condition to allow all child accounts to publish to SNS topic
In the existing solution AWS account numbers are hardcoded into the SNS topic (AccessKeyRotationTopic) access policy in the parent account. If a new AWS account is added to the organization it will not be able to publish to the AccessKeyRotationTopic due to access policy restriction. 

The new solution allows all the child accounts belonging to a given organization to publish to the SNS topic (AccessKeyRotationTopic). It also holds good for newly added accounts to AWS organization.